[router] merge link into single component
